updated the folder structure and added the root import file which exports all the components
