+++ src/RootImport.js
@@ -1,0 +1,14 @@
+export { default as Error } from "./Sections/Error";
+export { default as Nav } from "./components/navbar/Nav";
+export { default as Home } from "./Sections/Home";
+export { default as Auth } from "./Sections/Auth";
+export { default as PrivateRoute } from "./Sections/user/PrivateRoute";
+export { default as Dashboard } from "./Sections/user/Dashboard";
+export { default as Profile } from "./Sections/user/Profile";
+export { default as Role } from "./Sections/user/Role";
+export { default as Location } from "./Sections/user/Location";
+export { default as RequestWaste } from "./components/waste/RequestWaste";
+export { default as AddWaste } from "./components/waste/AddWaste";
+export { default as setUser } from "./store/slice/userSlice";
+export { default as WasteProductCard } from "./components/home/WasteProductCard";
+export { default as UserCard } from "./components/home/UserCard";
