--- src/Sections/FakeWaste.jsx
@@ -1,84 +1,0 @@
-const fakewastes = [
-    {
-        "category": "Food",
-        "measuringUnit": "Kg",
-        "amount": 5,
-        "price": 3.99,
-        "departue": "2023-10-20T10:00:00Z",
-        "recycled": false
-    },
-    {
-        "category": "E-waste",
-        "measuringUnit": "pcs",
-        "amount": 20,
-        "price": 15.99,
-        "departue": "2023-10-21T14:30:00Z",
-        "recycled": true
-    },
-    {
-        "category": "plastic-waste",
-        "measuringUnit": "Kg",
-        "amount": 8,
-        "price": 2.49,
-        "departue": "2023-10-22T08:45:00Z",
-        "recycled": true
-    },
-    {
-        "category": "Food",
-        "measuringUnit": "Kg",
-        "amount": 15,
-        "price": 6.99,
-        "departue": "2023-10-23T11:15:00Z",
-        "recycled": false
-    },
-    {
-        "category": "E-waste",
-        "measuringUnit": "pcs",
-        "amount": 30,
-        "price": 25.99,
-        "departue": "2023-10-24T09:20:00Z",
-        "recycled": true
-    },
-    {
-        "category": "plastic-waste",
-        "measuringUnit": "Kg",
-        "amount": 12,
-        "price": 4.49,
-        "departue": "2023-10-25T13:00:00Z",
-        "recycled": false
-    },
-    {
-        "category": "Food",
-        "measuringUnit": "Kg",
-        "amount": 7,
-        "price": 5.99,
-        "departue": "2023-10-26T16:45:00Z",
-        "recycled": true
-    },
-    {
-        "category": "E-waste",
-        "measuringUnit": "pcs",
-        "amount": 25,
-        "price": 19.99,
-        "departue": "2023-10-27T07:30:00Z",
-        "recycled": true
-    },
-    {
-        "category": "plastic-waste",
-        "measuringUnit": "Kg",
-        "amount": 9,
-        "price": 3.99,
-        "departue": "2023-10-28T14:00:00Z",
-        "recycled": false
-    },
-    {
-        "category": "Food",
-        "measuringUnit": "Kg",
-        "amount": 10,
-        "price": 7.99,
-        "departue": "2023-10-29T10:30:00Z",
-        "recycled": true
-    }
-];
-
-export default fakewastes;
