--- src/Sections/Waste.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import { Outlet } from 'react-router-dom'
-
-const Waste = () => {
-
-    return (
-        <div>
-            <Outlet/>
-        </div>
-    )
-}
-
-export default Waste
