+++ src/Sections/user/FakeData.jsx
@@ -1,0 +1,94 @@
+const users =  [
+    {
+        "email": "user1@example.com",
+        "password": "password1",
+        "userName": "User One",
+        "pfp": "profile_pic1.jpg",
+        "contact": [1234567890],
+        "role": "Supplier",
+        "location": "Location 1"
+    },
+    {
+        "email": "user2@example.com",
+        "password": "password2",
+        "userName": "User Two",
+        "pfp": "profile_pic2.jpg",
+        "contact": [9876543210],
+        "role": "Receiver",
+        "location": "Location 2"
+    },
+    {
+        "email": "user3@example.com",
+        "password": "password3",
+        "userName": "User Three",
+        "pfp": "profile_pic3.jpg",
+        "contact": [5555555555],
+        "role": "Supplier",
+        "location": "Location 3"
+    },
+    {
+        "email": "user4@example.com",
+        "password": "password4",
+        "userName": "User Four",
+        "pfp": "profile_pic4.jpg",
+        "contact": [7777777777],
+        "role": "Receiver",
+        "location": "Location 4"
+    },
+    {
+        "email": "user5@example.com",
+        "password": "password5",
+        "userName": "User Five",
+        "pfp": "profile_pic5.jpg",
+        "contact": [1111111111],
+        "role": "Supplier",
+        "location": "Location 5"
+    },
+    {
+        "email": "user6@example.com",
+        "password": "password6",
+        "userName": "User Six",
+        "pfp": "profile_pic6.jpg",
+        "contact": [9999999999],
+        "role": "Supplier",
+        "location": "Location 6"
+    },
+    {
+        "email": "user7@example.com",
+        "password": "password7",
+        "userName": "User Seven",
+        "pfp": "profile_pic7.jpg",
+        "contact": [2222222222],
+        "role": "Receiver",
+        "location": "Location 7"
+    },
+    {
+        "email": "user8@example.com",
+        "password": "password8",
+        "userName": "User Eight",
+        "pfp": "profile_pic8.jpg",
+        "contact": [8888888888],
+        "role": "Supplier",
+        "location": "Location 8"
+    },
+    {
+        "email": "user9@example.com",
+        "password": "password9",
+        "userName": "User Nine",
+        "pfp": "profile_pic9.jpg",
+        "contact": [3333333333],
+        "role": "Receiver",
+        "location": "Location 9"
+    },
+    {
+        "email": "user10@example.com",
+        "password": "password10",
+        "userName": "User Ten",
+        "pfp": "profile_pic10.jpg",
+        "contact": [6666666666],
+        "role": "Supplier",
+        "location": "Location 10"
+    }
+]
+
+export default users;
